Move console interception to context
